Omit abort on conflict skipping, if no cherry-pick is in progress
Checks whether any reverts or cherry-picks are in progress,
when user answers Yes to the "Continue from the next commit (Yes/no)?"
prompt when a the process runs into a merge conflict.

If there is no active revery or cherry-pick in progress, expect
that the conflict has been resolved by the user and continue
to the next commit without abort.

Previous, answering Yes, would invoke abort command, which would
fail if the user had already resolved conflicts. The previous workflow
expected that the user would answer no, and then run the git-draft
command again after the conflict had been resolved.

Now with this change, running the command again isn't necessary,
and it's enough to keep git-draft open on the background.

Closes #6